Simplify multi-namespace controller setup
Simplify the multi-namespace controller setup by leveraging the source.Informer
and cache.Cache structures. This allows the deletion of the OperatorManager
structure and concept.

+++ hack/uninstall.sh
@@ -1,18 +1,26 @@
 #!/bin/bash
 set -uo pipefail
 
+OPERAND_ONLY="${OPERAND_ONLY:-}"
+
 # Disable the CVO
 oc scale --replicas 0 -n openshift-cluster-version deployments/cluster-version-operator
 
 # Uninstall the cluster-ingress-operator
 oc delete -n openshift-ingress-operator deployments/ingress-operator
 oc patch -n openshift-ingress-operator clusteringresses/default --patch '{"metadata":{"finalizers": []}}' --type=merge
-oc delete -n openshift-ingress-operator clusteroperators.operatorstatus.openshift.io/openshift-ingress
+oc delete clusteroperator.config.openshift.io/openshift-ingress-operator
 oc delete --force --grace-period=0 -n openshift-ingress-operator clusteringresses/default
-oc delete namespaces/openshift-ingress-operator
+
+if [ "$OPERAND_ONLY" != "true" ]; then
+  oc delete namespaces/openshift-ingress-operator
+fi
 oc delete namespaces/openshift-ingress
-oc delete clusterroles/openshift-ingress-operator
-oc delete clusterroles/openshift-ingress-router
-oc delete clusterrolebindings/openshift-ingress-operator
-oc delete clusterrolebindings/openshift-ingress-router
-oc delete customresourcedefinition.apiextensions.k8s.io/clusteringresses.ingress.openshift.io
+
+if [ "$OPERAND_ONLY" != "true" ]; then
+  oc delete clusterroles/openshift-ingress-operator
+  oc delete clusterroles/openshift-ingress-router
+  oc delete clusterrolebindings/openshift-ingress-operator
+  oc delete clusterrolebindings/openshift-ingress-router
+  oc delete customresourcedefinition.apiextensions.k8s.io/clusteringresses.ingress.openshift.io
+fi
